Cleanup schema test for kernel lifetimes XML
The test was originally meant to be an early indicator of malformed changes to kernel-lifetimes.xml file, but the test is unable to serve this purpose and is simply a maintenance burden. Moreover, the kernel end-of-life check which consumes kernel lifetimes XML file is planned to be removed from VTS test suite and therefore this test is no longer required.

Bug: 425765063
Change-Id: If0169c92551d0516161303ebba21418ae16746d4